feat(): show collection when save

+++ lib/features/post/presentation/widgets/collection/collection_more_button.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+import '../../../../../core/di/injection.dart';
 import '../../provider/collection_cubit.dart';
 import '../../../../../core/widgets/custom_bottom_sheet.dart';
 import '../../../data/models/collection_model.dart';
@@ -16,7 +17,7 @@ class CollectionMoreButton extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final cubit = context.read<CollectionCubit>();
+    final cubit = getIt<CollectionCubit>();
 
     return IconButton(
       icon: Icon(Icons.more_vert, color: iconColor),
@@ -31,8 +32,44 @@ class CollectionMoreButton extends StatelessWidget {
             icon: const Icon(Icons.delete, color: Colors.red),
             title: 'Xóa bộ sưu tập',
             onTap: () async {
-              Navigator.pop(context);
-              await cubit.deleteCollection(collection.id);
+              final confirm = await showDialog<bool>(
+                context: context,
+                builder: (_) => AlertDialog(
+                  title: const Text('Xác nhận xóa'),
+                  content: Text.rich(
+                    TextSpan(
+                      text: 'Bạn có chắc muốn xóa bộ sưu tập ',
+                      style: const TextStyle(fontSize: 16),
+                      children: [
+                        TextSpan(
+                          text: collection.name,
+                          style: const TextStyle(
+                            fontWeight: FontWeight.bold,
+                            color: Colors.redAccent,
+                          ),
+                        ),
+                        const TextSpan(text: ' không?'),
+                      ],
+                    ),
+                  ),
+                  actions: [
+                    TextButton(
+                      onPressed: () => Navigator.pop(context, false),
+                      child: const Text('Hủy'),
+                    ),
+                    ElevatedButton(
+                      onPressed: () => Navigator.pop(context, true),
+                      style: ElevatedButton.styleFrom(backgroundColor: Colors.red),
+                      child: const Text('Xóa'),
+                    ),
+                  ],
+                ),
+              );
+
+              if (confirm == true) {
+                await cubit.deleteCollection(collection.id);
+                if (context.mounted) Navigator.pop(context); ////
+              }
             },
           ),
         ]);
